requirements.txt
watchdog
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
schedule
requests
python-dotenv
